Renames Router to BaseRouter and ReactRouter to Router
