+++ index.js
@@ -1,7 +1,7 @@
 import {Schema, Route} from './Common';
 import TabBar from './TabBar';
 import Actions from './Actions';
-import Router from './ReactRouter';
+import Router from './Router';
 import Animations from './Animations';
 
 module.exports = {Schema, Route, TabBar, Actions, Router, Animations};
